Don't overwrite existing files

+++ AFQ/api/group.py
@@ -208,8 +208,9 @@ class GroupAFQ(object):
 
         pl_desc_file = op.join(self.afq_path, 'dataset_description.json')
 
-        with open(pl_desc_file, 'w') as outfile:
-            json.dump(pipeline_description, outfile)
+        if not op.exists(pl_desc_file):
+            with open(pl_desc_file, 'w') as outfile:
+                json.dump(pipeline_description, outfile)
 
         self.subjects = bids_layout.get(return_type='id', target='subject')
         if not len(self.subjects):
@@ -410,7 +411,12 @@ class GroupAFQ(object):
             self.afq_path, "tract_profiles.csv"))
         os.makedirs(op.dirname(out_file), exist_ok=True)
         _df = clean_pandas_df(_df)
-        _df.to_csv(out_file, index=False)
+        try:
+            _df.to_csv(out_file, index=False)
+        except:
+            logger.warning((
+                "Unable to update combined tract profile. "
+                "This is likely due to file permissions."))
         return _df
 
     def get_streamlines_json(self):
@@ -777,7 +783,8 @@ class GroupAFQ(object):
                 self.afq_path,
                 (f"bundle-{bundle_name}_view-{view}"
                     f"_idx-{curr_file_num}_montage.png")))
-            curr_img.save(save_path)
+            if not op.exists(save_path):
+                curr_img.save(save_path)
             all_fnames.append(save_path)
 
         this_img_trimmed = {}
